moved submitted tests to appropriate test class
